+++ packages/blade/src/components/ActionList/ActionList.tsx
@@ -7,9 +7,9 @@ import { ActionListBox } from './ActionListBox';
 import { componentIds } from './componentIds';
 import { useDropdown } from '~components/Dropdown/useDropdown';
 import { assignWithoutSideEffects, makeAccessible, metaAttribute, MetaConstants } from '~utils';
-import { useTheme } from '~components/BladeProvider';
 import { useBottomSheetContext } from '~components/BottomSheet/BottomSheetContext';
 import type { TestID } from '~src/_helpers/types';
+import type { SurfaceLevels } from '~tokens/theme/theme';
 
 type ActionListContextProp = Pick<ActionListProps, 'surfaceLevel'>;
 const ActionListContext = React.createContext<ActionListContextProp>({ surfaceLevel: 2 });
@@ -29,7 +29,7 @@ type ActionListProps = {
   /**
    * Decides the backgroundColor of ActionList
    */
-  surfaceLevel?: 2 | 3;
+  surfaceLevel?: Exclude<SurfaceLevels, 1>;
 } & TestID;
 
 /**
@@ -82,30 +82,21 @@ const _ActionList = ({ children, surfaceLevel = 2, testID }: ActionListProps): J
     actionListItemRef,
     selectionType,
     dropdownBaseId,
-    setSelectedIndices,
     dropdownTriggerer,
     hasFooterAction,
   } = useDropdown();
 
-  const { theme } = useTheme();
   const { isInBottomSheet } = useBottomSheetContext();
 
-  const {
-    sectionData,
-    childrenWithId,
-    actionListOptions,
-    defaultSelectedIndices,
-  } = React.useMemo(() => getActionListProperties(children), [children]);
+  const { sectionData, childrenWithId, actionListOptions } = React.useMemo(
+    () => getActionListProperties(children),
+    [children],
+  );
 
   React.useEffect(() => {
     setOptions(actionListOptions);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [actionListOptions]);
-
-  React.useEffect(() => {
-    setSelectedIndices(defaultSelectedIndices);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   const actionListContainerRole = getActionListContainerRole(hasFooterAction, dropdownTriggerer);
   const actionListItemWrapperRole = getActionListItemWrapperRole(
@@ -136,7 +127,6 @@ const _ActionList = ({ children, surfaceLevel = 2, testID }: ActionListProps): J
         <StyledActionList
           isInBottomSheet={isInBottomSheet}
           surfaceLevel={surfaceLevel}
-          elevation={isInBottomSheet ? undefined : theme.shadows.androidElevation.level[2]}
           id={`${dropdownBaseId}-actionlist`}
           {...makeAccessible({
             role: actionListContainerRole,
@@ -160,8 +150,7 @@ const _ActionList = ({ children, surfaceLevel = 2, testID }: ActionListProps): J
 };
 
 const ActionList = assignWithoutSideEffects(React.memo(_ActionList), {
-  componentId: componentIds.ActionList,
-  displayName: 'ActionList',
+  displayName: componentIds.ActionList,
 });
 
 export { ActionList, useActionListContext, ActionListProps };