+++ packages/blade/src/components/ActionList/actionListUtils.ts
@@ -55,12 +55,10 @@ const getActionListProperties = (
   sectionData: SectionData;
   childrenWithId?: React.ReactNode[] | null;
   actionListOptions: OptionsType;
-  defaultSelectedIndices: number[];
 } => {
   const sectionData: SectionData = [];
   let currentSection: string | null = null;
   const actionListOptions: OptionsType = [];
-  const defaultSelectedIndices: number[] = [];
 
   const getActionListItemWithId = (
     child: React.ReactNode,
@@ -70,7 +68,21 @@ const getActionListProperties = (
       actionListOptions.push({
         title: child.props.title,
         value: child.props.value,
-        href: child.props.href,
+        onClickTrigger: (value) => {
+          const anchorLink = child.props.href;
+          child.props.onClick?.({
+            name: child.props.value,
+            value: child.props.isSelected ?? value,
+          });
+
+          if (anchorLink && !isReactNative()) {
+            const target = child.props.target ?? '_self';
+            window.open(anchorLink, target);
+            if (window.top) {
+              window.top.open(anchorLink, target);
+            }
+          }
+        },
       });
       const currentIndex = actionListOptions.length - 1;
 
@@ -93,10 +105,6 @@ const getActionListProperties = (
             },
           ],
         });
-      }
-
-      if (child.props.isDefaultSelected) {
-        defaultSelectedIndices.push(currentIndex);
       }
 
       const clonedChild = React.cloneElement(child, {
@@ -160,7 +168,6 @@ const getActionListProperties = (
     sectionData,
     childrenWithId,
     actionListOptions,
-    defaultSelectedIndices,
   };
 };