+++ packages/blade/src/components/Amount/__tests__/Amount.web.test.tsx
@@ -23,7 +23,7 @@ describe('<Amount />', () => {
   it('should throw an error when a string is passed', () => {
     const mockConsoleError = jest.spyOn(console, 'error').mockImplementation();
     // @ts-expect-error testing failure case when value is passed as a string
-    expect(() => renderWithTheme(<Amount value={'10000'} />)).toThrow(
+    expect(() => renderWithTheme(<Amount value="10000" />)).toThrow(
       '[Blade: Amount]: `value` prop must be of type `number` for Amount.',
     );
     mockConsoleError.mockRestore();