+++ packages/blade/src/components/Checkbox/useCheckbox.ts
@@ -18,6 +18,7 @@ type UseCheckboxProps = Pick<
   | 'name'
   | 'value'
 > & {
+  role?: 'checkbox' | 'switch';
   hasError?: boolean;
   hasHelperText?: boolean;
 };
@@ -35,6 +36,7 @@ function setMixed(element: HTMLInputElement, mixed?: boolean) {
 }
 
 const useCheckbox = ({
+  role = 'checkbox',
   isChecked,
   defaultChecked,
   isIndeterminate,
@@ -92,7 +94,7 @@ const useCheckbox = ({
   const { inputId, errorTextId, helpTextId } = useFormId('checkbox');
 
   const accessibilityProps = makeAccessible({
-    role: 'checkbox',
+    role,
     required: Boolean(isRequired),
     invalid: Boolean(hasError),
     disabled: Boolean(isDisabled),